+++ CHANGES.rst
@@ -17,6 +17,13 @@ utilipy.tests.helper
   + phase-wrap support
 
 
+utilipy.utils.pickle
+^^^^^^^^^^^^^^^^^^^^
+
+- Added configuration for using dill over pickle [#26]
+  Changed default from True to False
+
+
 API Changes
 -----------
 
@@ -31,6 +38,9 @@ API Changes
 
 - Module ``ipython`` is not imported to top-level namespace [#20]
   because this is not always used in an ipython environment.
+
+
+- Changed default from True to False in dill-over-pickle config [#26]
 
 
 Bug Fixes